+++ test/open_weather_map_client_test.dart
@@ -3,5 +3,11 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:open_weather_map_client/open_weather_map_client.dart';
 
 void main() {
-  test('adds one to input values', () {});
+  final OpenWeatherMap _api = OpenWeatherMap(
+    apiKey: '856822fd8e22db5e1ba48c0e7d69844a',
+  );
+  test('Check values from Weather class', () async {
+    //final Weather weather = _api.currentWeather;
+    //expect(weather.humidity, isA<int>());
+  });
 }