+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.1.0
+* Update readme.
+* Update example with provider.
+* Use provider package for management the state.
+
+## 0.0.9
+* Update example without provider.
+
+## 0.0.8
+* Now you can get temperature in Celsius and Fahrenheit
+
+## 0.0.7
+* Functional version.
+
+## 0.0.6
+* Testing version without Provider package
+
+## 0.0.5
+* Testing version with Provider package
+
+## 0.0.4
+* Add current weather get.
+
 ## 0.0.3
 * Add GitHub repository link.
 * Add city model.